Add: Initial Solution to Raindrops
This is a verbose and primitive solution, which passes test, but
needs to be refactored.

The algorith is reasonable, but the same approach could be
implemented without separate functions, and the string building
could be more elegant.

However, I will implement this a different way, as it provides
an opportunity to use some of the stream features of newer Javas.